Fix formatting, cmd+a
